requirements.txt
django
djangorestframework
django-filter
pyyaml
uritemplate